added asterisk to words
When we wrote this, we thought if important to write out the words "f*g" and "n*gger" without asterisks on the grounds that we were referring to them, not using them, and we wanted the impact of them to be stark. In the intervening years, I'm not comfortable even referring to these words without asterisks. So I added the asterisks.  — DW